--- __tests__/resolvers.test.js
@@ -1,3 +1,0 @@
-test('a sample test', () => {
-	expect(2).toBe(2);
-});
