+++ graphql/types/__tests__/categories.test.js
